Use sync client for setup / teardown ; Ensure sync / async connection is terminated correctly

+++ integrations/mongodb_atlas/tests/test_document_store_async.py
@@ -35,25 +35,44 @@ def test_init_is_lazy(_mock_client):
 class TestDocumentStoreAsync(FilterableDocsFixtureMixin):
 
     @pytest.fixture
-    def document_store(self):
+    async def document_store(self):
         database_name = "haystack_integration_test"
         collection_name = "test_collection_" + str(uuid4())
         connection_string = os.environ["MONGO_CONNECTION_STRING"]
-        with MongoClient(connection_string, driver=DriverInfo(name="MongoDBAtlasHaystackIntegration")) as connection:
-            database = connection[database_name]
+
+        # We're using the sync client for setup/teardown ease
+        sync_client = MongoClient(connection_string, driver=DriverInfo(name="MongoDBAtlasHaystackIntegration"))
+
+        store = MongoDBAtlasDocumentStore(
+            database_name=database_name,
+            collection_name=collection_name,
+            vector_search_index="cosine_index",
+            full_text_search_index="full_text_index",
+        )
+        try:
+            database = sync_client[database_name]
             if collection_name in database.list_collection_names():
                 database[collection_name].drop()
             database.create_collection(collection_name)
             database[collection_name].create_index("id", unique=True)
 
-            store = MongoDBAtlasDocumentStore(
-                database_name=database_name,
-                collection_name=collection_name,
-                vector_search_index="cosine_index",
-                full_text_search_index="full_text_index",
-            )
+            # Initialize the async connection before yielding
+            await store._ensure_connection_setup_async()
+
             yield store
-            database[collection_name].drop()
+        finally:
+            # Ensure async connection is closed before synchronous teardown
+            if store._connection_async:
+                await store.connection.close()
+
+            # Synchronous teardown
+            if sync_client:
+                try:
+                    database = sync_client[database_name]
+                    if collection_name in database.list_collection_names():
+                        database[collection_name].drop()
+                finally:
+                    sync_client.close()
 
     async def test_write_documents_async(self, document_store: MongoDBAtlasDocumentStore):
         docs = [Document(content="some text")]
